Adding custom colors
